backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
numpy
python-dateutil
google-generativeai
python-dotenv
